# api/openapi.yaml
openapi: 3.0.3
info:
  title: FolioPath HTTP API
  version: 0.1.0
  description: |
    Authoritative HTTP contract through the FolioPath POST-MVP-5 scope,
    revision 2. Revision-1 image semantic operations are contract-frozen.
    Revision-2 controlled tag suggestions, video semantic search, and local
    anonymous face/person curation remain gated from production composition
    until INT-S1R2 is accepted.

    All business endpoints use an authenticated same-origin browser session.
    State-changing authenticated operations additionally require the
    session-bound `X-CSRF-Token`. First-run setup and login do not yet have a
    session, so they require a validated same-origin `Origin` header.

    IDs and cursors are opaque. Media content endpoints accept only resource
    IDs, never filesystem paths. Management responses may expose a library
    root relative to the allowed `/library` container mount and a server-built
    `/library/...` display label; no response may expose a host path.
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
...
servers:
  - url: /
    description: Same-origin FolioPath server
security:
  - cookieAuth: []
...
paths:
...
  /api/v1/libraries/{libraryId}/ai/semantic:
    parameters:
      - $ref: '#/components/parameters/LibraryIDParameter'
    get:
      tags: [Semantic search]
      operationId: getLibrarySemanticSettings
      summary: Get one library's semantic-search settings and coverage
      x-requirements: [FR-INT-001, FR-INT-015, FR-INT-016]
      x-error-codes:
        '401': [authentication_required, session_expired]
        '404': [library_not_found]
        '429': [rate_limited]
        '500': [internal_error]
      responses:
        '200':
          description: Semantic settings and truthful coverage snapshot.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SemanticLibrarySettings'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [Semantic search]
      operationId: updateLibrarySemanticSettings
      summary: Enable or disable semantic analysis for one library
      security:
        - cookieAuth: []
          csrfToken: []
      x-requirements: [FR-INT-001, FR-INT-016, NFR-INT-003]
      x-error-codes:
        '400': [invalid_request]
        '401': [authentication_required, session_expired]
        '403': [csrf_invalid]
        '404': [library_not_found]
        '409': [model_unavailable]
        '412': [precondition_failed]
        '422': [validation_failed]
        '428': [precondition_required]
        '429': [rate_limited]
        '500': [internal_error]
      parameters:
        - $ref: '#/components/parameters/IfMatchHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SemanticLibrarySettingsUpdate'
      responses:
        '200':
          description: Updated semantic settings.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SemanticLibrarySettings'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/libraries/{libraryId}/ai/semantic/jobs:
    parameters:
      - $ref: '#/components/parameters/LibraryIDParameter'
    post:
      tags: [Semantic search]
      operationId: requestLibrarySemanticJob
      summary: Request missing-only or full semantic generation work
      security:
        - cookieAuth: []
          csrfToken: []
      x-requirements: [FR-INT-005, FR-INT-015, FR-INT-016, NFR-INT-002]
      x-error-codes:
        '400': [invalid_request]
        '401': [authentication_required, session_expired]
        '403': [csrf_invalid]
        '404': [library_not_found]
        '409': [ai_disabled, model_unavailable, idempotency_conflict]
        '429': [rate_limited]
        '500': [internal_error]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SemanticJobRequest'
      responses:
        '202':
          description: A bounded semantic job was created or an equivalent active intent was coalesced.
          headers:
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
            ETag:
              $ref: '#/components/headers/ETag'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIModelOperation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/libraries/{libraryId}/ai/semantic/clear:
    parameters:
      - $ref: '#/components/parameters/LibraryIDParameter'
    post:
      tags: [Semantic search]
      operationId: clearLibrarySemanticData
      summary: Clear only rebuildable semantic data for one library
      security:
        - cookieAuth: []
          csrfToken: []
      x-requirements: [FR-INT-001, FR-INT-005, NFR-INT-003]
      x-error-codes:
        '400': [invalid_request]
        '401': [authentication_required, session_expired]
        '403': [csrf_invalid]
        '404': [library_not_found]
        '409': [idempotency_conflict]
        '412': [precondition_failed]
        '422': [validation_failed]
        '428': [precondition_required]
        '429': [rate_limited]
        '500': [internal_error]
      parameters:
        - $ref: '#/components/parameters/IfMatchHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClearSemanticDataRequest'
      responses:
        '202':
          description: Restart-safe bounded deletion operation accepted.
          headers:
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
            ETag:
              $ref: '#/components/headers/ETag'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIModelOperation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/semantic/assets:
    get:
      tags: [Semantic search]
      operationId: searchAssetsBySemantics
      summary: Search indexed images by Chinese or English visual description
      description: |
        Omit `libraryId` for all enabled libraries. `directoryId` requires a
        library and selects direct children unless `recursive=true`. Results
        use `score DESC, asset_id ASC`; the score is not exposed as a percent.
        The cursor binds normalized query fingerprint, scope, catalog revision,
        model generation, semantic generation, progress revision, and the last
        ordering tuple. Query text is never logged or placed in the cursor.
      x-requirements: [FR-INT-002, FR-INT-003, FR-INT-005, FR-INT-016, NFR-INT-001, NFR-INT-004, NFR-INT-005]
      x-error-codes:
        '400': [invalid_request, invalid_cursor]
        '401': [authentication_required, session_expired]
        '409': [ai_disabled, semantic_not_ready, semantic_cursor_stale]
        '429': [rate_limited, semantic_busy]
        '503': [model_unavailable]
        '500': [internal_error]
      parameters:
        - $ref: '#/components/parameters/SemanticQueryParameter'
        - $ref: '#/components/parameters/OptionalLibraryIDParameter'
        - $ref: '#/components/parameters/DirectoryFilterParameter'
        - $ref: '#/components/parameters/RecursiveParameter'
        - $ref: '#/components/parameters/CursorParameter'
        - $ref: '#/components/parameters/LimitParameter'
      responses:
        '200':
          description: Stable image-only semantic result page with truthful coverage.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SemanticAssetPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: foliopath_session
      description: |
        High-entropy, opaque session cookie backed by a revocable server-side
        session. It is HttpOnly and SameSite=Strict; deployments using HTTPS
        set Secure. The database stores no replayable plaintext token.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        Session-bound token returned by setup, login, and session responses.
        It is required together with `cookieAuth` on authenticated
        state-changing operations. It must never be placed in a URL.
  parameters:
...
    LibraryIDParameter:
      name: libraryId
      in: path
      required: true
      description: Opaque media-library ID. It contains no path information.
      schema:
        $ref: '#/components/schemas/ResourceID'
...
    IfMatchHeader:
      name: If-Match
      in: header
      required: true
      description: Strong ETag from the latest GET. Missing values fail with 428 and stale values with 412.
      schema:
        type: string
        minLength: 3
        maxLength: 256
...
  headers:
...
    RequestID:
      description: Server-generated opaque correlation ID. Client-supplied values are not trusted.
      schema:
        type: string
        pattern: '^req_[A-Za-z0-9_-]+$'
        maxLength: 128
...
    ETag:
      description: |
        Strong opaque representation validator. For a Library it changes when
        any represented field changes, including scan-derived status/counts.
      schema:
        type: string
        example: '"fp-8b51cdd5"'
...